+++ src/ui/molecules/SearchSection/Salon/Salon.tsx
@@ -1,0 +1,6 @@
+
+export const Salon = () => {
+    return (
+        <div>Salon</div>
+    )
+}
