+++ src/ui/molecules/SearchSection/Salon/index.ts
@@ -1,0 +1,1 @@
+export * from './Salon';
