+++ src/ui/molecules/SearchSection/index.ts
@@ -1,1 +1,4 @@
+export * from './Salon';
 export * from './SearchSection';
+export * from './SelectionSection';
+export * from './Services';
